RenderPass now has a size so that the viewport can be set correctly. Remove threading from FBX loading to stop I/O thrashing. Add some notes to the README on code design.

+++ data/shaders/texturedVertLit.frag
@@ -9,7 +9,7 @@ out vec4 outputColor;
 uniform vec3 lightDirection;
 uniform vec3 lightColor;
 uniform vec3 cameraPosition;
-uniform float specularPower = 1;
+uniform int specularPower = 1;
 uniform sampler2D diffuseSampler;
 
 void main()
